backend/requirements.txt
openai==1.57.0
fastapi==0.115.6
sqlalchemy==2.0.34
groq==0.13.1
exa-py==1.7.1
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10